Split VT emulator into "text only" and Qt dependant parts
