+++ main/src/emulator/Term.jest.ts
@@ -4,15 +4,15 @@
  * This source code is licensed under the MIT license which is detailed in the LICENSE.txt file.
  */
 import { performance } from "node:perf_hooks";
-import {Emulator, MAX_WRITE_BUFFER_SIZE, Platform} from "./Term.js";
-import {RenderEvent, Line, MinimalKeyboardEvent, DataEvent} from "term-api";
+import {TextEmulator, MAX_WRITE_BUFFER_SIZE, Platform} from "./TextTerm.js";
+import {RenderEvent, Line, MinimalKeyboardEvent, DataEvent} from "text-term-api";
 import { STYLE_MASK_CURSOR } from "extraterm-char-cell-line";
 import fs from "node:fs";
 
 
 test("basic", done => {
   async function testBasic(): Promise<void> {
-    const emulator = new Emulator({platform: <Platform> process.platform,
+    const emulator = new TextEmulator({platform: <Platform> process.platform,
       performanceNowFunc: () => performance.now()});
     emulator.write('Hello');
 
@@ -27,7 +27,7 @@ test("basic", done => {
 
 test("wrap", done => {
   async function testWrap(): Promise<void> {
-    const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+    const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
       performanceNowFunc: () => performance.now()});
     emulator.write('abcdefghijklmnopqrstuvwxyz');
 
@@ -42,7 +42,7 @@ test("wrap", done => {
 
 test("scroll one", done => {
   async function testScrollOne(): Promise<void> {
-    const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+    const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
       performanceNowFunc: () => performance.now()});
 
     emulator.write('1\n');
@@ -67,7 +67,7 @@ test("scroll one", done => {
 
 test("render device", done => {
   async function testRenderDevice(): Promise<void> {
-    const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+    const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
       performanceNowFunc: () => performance.now()});
     const device = new RenderDevice();
     emulator.onRender(device.renderEventHandler.bind(device));
@@ -93,7 +93,7 @@ test("render device", done => {
 
 test("move cursor", done => {
   async function testMoveCursor(): Promise<void> {
-    const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+    const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
       performanceNowFunc: () => performance.now()});
     const device = new RenderDevice();
     emulator.onRender(device.renderEventHandler.bind(device));
@@ -123,7 +123,7 @@ describe.each([
 
   test(`UTF-16 surrogate handling: ${name}`, done => {
     async function testSurrogateHandling(): Promise<void> {
-      const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+      const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
         performanceNowFunc: () => performance.now()});
       const device = new RenderDevice();
       emulator.onRender(device.renderEventHandler.bind(device));
@@ -144,7 +144,7 @@ describe.each([
 
 test("Move rows above cursor to scrollback", done => {
   async function testMoveRowsAboveCursorToScrollback(): Promise<void> {
-    const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+    const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
       performanceNowFunc: () => performance.now()});
     const device = new RenderDevice();
     emulator.onRender(device.renderEventHandler.bind(device));
@@ -180,7 +180,7 @@ test("Move rows above cursor to scrollback", done => {
 
 test("hyperlink", done => {
   async function testHyperlink(): Promise<void> {
-    const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+    const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
       performanceNowFunc: () => performance.now()});
     const device = new RenderDevice();
     emulator.onRender(device.renderEventHandler.bind(device));
@@ -202,7 +202,7 @@ describe.each([
 ])("Keyboard input keyPress()", (ev: MinimalKeyboardEvent, output: string) => {
 
   test(`${JSON.stringify(ev)} => ${JSON.stringify(output)}`, done => {
-    const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+    const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
       performanceNowFunc: () => performance.now()});
     let collectedData = "";
     emulator.onData( (event: DataEvent): void => {
@@ -249,7 +249,7 @@ describe.each([
 ])("Keyboard input keyDown()", (ev: MinimalKeyboardEvent, output: string) => {
 
   test(`${JSON.stringify(ev)} => ${JSON.stringify(output)}`, done => {
-    const emulator = new Emulator({platform: <Platform> process.platform, rows: 10, columns: 20,
+    const emulator = new TextEmulator({platform: <Platform> process.platform, rows: 10, columns: 20,
       performanceNowFunc: () => performance.now() });
     let collectedData = "";
     emulator.onData( (event: DataEvent): void => {
@@ -269,7 +269,7 @@ function waitOnEmulator(emulator: Emulator): Promise<void> {
   });
 }
 
-async function waitOnEmulator(emulator: Emulator): Promise<void> {
+async function waitOnEmulator(emulator: TextEmulator): Promise<void> {
   let done = false;
   if (emulator.getWriteBufferStatus().bufferSize === MAX_WRITE_BUFFER_SIZE) {
     return;
@@ -285,7 +285,7 @@ async function waitOnEmulator(emulator: Emulator): Promise<void> {
   }
 }
 
-function readEmulatorScreenString(emulator: Emulator): string {
+function readEmulatorScreenString(emulator: TextEmulator): string {
   let result = lineToString(emulator.lineAtRow(0));
   let row = 1;
