+++ packages/term-api-lineimpl/src/main.ts
@@ -1,4 +1,4 @@
 /**
  * Copyright 2020 Simon Edwards <simon@simonzone.com>
  */
-export { LineImpl, CellWithHyperlink } from "./LineImpl.js";
+export { LineImpl } from "./LineImpl.js";
