+++ packages/text-term-api-lineimpl/.eslintrc.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+	extends: "extraterm"
+};
