+++ packages/text-term-api-lineimpl/README.md
@@ -1,0 +1,1 @@
+Shareable implementation of TermApi.Line API. This mainly adds link related methods.
