+++ packages/text-term-api-lineimpl/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "term-api-lineimpl",
+  "name": "text-term-api-lineimpl",
   "version": "1.0.0",
   "description": "Implementation of TermAPI's Line interface",
   "author": "Simon Edwards",
@@ -13,11 +13,9 @@
   },
   "dependencies": {
     "extraterm-char-cell-line": "1.0.0",
-    "extraterm-char-render-canvas": "1.0.0",
     "extraterm-data-structures": "1.0.0",
     "extraterm-unicode-utilities": "1.0.0",
-    "term-api": "1.0.0",
-    "text-term-api-lineimpl": "1.0.0"
+    "text-term-api": "1.0.0"
   },
   "devDependencies": {
     "@types/jest": "27.0.1",
@@ -34,6 +32,7 @@
     "build-fast": "tsc",
     "clean": "shx rm -rf dist",
     "lint": "eslint \"src/**/*.ts\"",
-    "lint-strict": "eslint --max-warnings 1 \"src/**/*.ts\""
+    "lint-strict": "eslint --max-warnings 1 \"src/**/*.ts\"",
+    "test": "yarn run build && yarn node --experimental-vm-modules $(yarn bin jest)"
   }
 }
