+++ packages/text-term-api/.eslintrc.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+	extends: "extraterm"
+};
