+++ packages/text-term-api/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../../tsconfig",
+    "compilerOptions": {
+        "outDir": "dist",
+        "rootDir": "src"
+    },
+    "include": [
+        "./src/**/*.ts"
+    ]
+}
